jwt网关配置

+++ mall-common/src/main/java/com/mall/common/pojo/PageResult.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 /**
  * program: spring-cloud-mall->PageResult
- * description: 返回分页结果
+ * description: 分页
  * author: gerry
- * created: 2019-11-23 10:49
+ * created: 2019-11-25 22:13
  **/
 public class PageResult<T> {
 
@@ -17,7 +17,7 @@ public class PageResult<T> {
     /**
      * 总页数
      */
-    private Long totalPage;
+    private Integer totalPage;
     /**
      * 当前页数据
      */
@@ -31,7 +31,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -53,11 +53,11 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public Long getTotalPage() {
+    public Integer getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(Integer totalPage) {
         this.totalPage = totalPage;
     }
 }
